requirements.txt
absl-py
accelerate==1.6.0
aiortc
av
diffusers
python-dotenv
fastapi
ffmpeg-python
google-genai
gradio
json_repair
librosa
llvmlite>=0.40.0
ml-collections
numba>=0.57.0
numpy
opencv-python
openai
pydub
pyngrok
scipy
streamlit
syncsdk
torch
tqdm
transformers==4.51.3
uvicorn[standard]
youtube-dl
fastvideo
pillow
requests
st_theme